pyproject.toml
[project]
name = "south-admin-django"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework-simplejwt>=5.5.1",
    "mysqlclient>=2.2.7",
]
